fix refresh redirect to signin
